+++ yash-builtin/CHANGELOG.md
@@ -20,6 +20,16 @@ A _private dependency_ is used internally and not visible to downstream users.
   invalid signal name is specified.
 - The `kill::Signal::to_name_and_number` method now returns `Cow<str>` instead
   of `yash_env::signal::Name`.
+- The `trap::syntax::interpret` function now takes one more argument,
+  `system: &S where S: yash_env::system::Signals`, to check the validity of
+  signal names and numbers.
+- The `trap::Command` now uses `yash_env::trap::Condition` instead of
+  `trap::CondSpec` to represent conditions.
+
+## Removed
+
+- The `trap::CondSpec` enum has been removed. Use `yash_env::trap::Condition`
+  instead.
 
 ## [0.14.0] - 2026-01-16